requirements.txt
Django==1.5.1
Markdown==2.3.1
PIL==1.1.7
South==0.7.6
argparse==1.2.1
astroid==1.0.0
decorator==3.4.0
distribute==0.6.24
django-bower==4.8
django-debug-toolbar==0.9.4
django-filter==0.7
djangorestframework==2.3.8
-e git+ssh://git@github.com/RickyCook/facebook-sdk.git@47f7ca3eb04e70c6cd4855ff4299bb89f104e397#egg=facebook_sdk-dev
httplib2==0.8
ipdb==0.8
ipython==1.1.0
jsonpath-rw==1.2.1
logilab-common==0.60.0
mimeparse==0.1.3
oauth2==1.5.211
oauthlib==0.6.0
pep8==1.4.6
ply==3.4
pylint==1.0.0
python-dateutil==2.1
python-openid==2.2.5
python-social-auth==0.1.14
requests==2.0.0
requests-oauthlib==0.4.0
simplejson==3.3.1
six==1.4.1
wsgiref==0.1.2
